Revert OTP21 fix and make examples easier to run (#28)
Revert OTP21 compilation fix changes - we will not be able to release it anyway (can not publish with overridden deps). Need to wait for new `amqp` version released. For a time being we are not able to compile `gen_rmq` on OTP21 - created issue for it - https://github.com/meltwater/gen_rmq/issues/29.