Consumer / Publisher non-blocking init (#71)
This sets up a Consumer / Publisher to be non-blocking during the init callback.

Related to issue #64.

There is a chance of creating a race condition with the current setup of doing `send(self(), :init)`. The new `GenServer.handle_continue/2` would be ideal for this situation. But that would require bumping the minimum elixir version to 1.7.