Remove Mix module from Publisher
Elixir 1.9 can package projects as releases (similar to Distillery). Releases do not include the Mix module.

A project using GenRMQ could not be run as a release because the GenRMQ Publisher uses Mix to lookup a default app_id. I've fixed this rather crudely by removing the code using Mix, tests still pass but it may be a breaking change for some apps.

It should be possible to conditionally compile the module with or without the Mix code depending on whether or not it is being compiled for release but I thought this would be confusing to end-users.