Connection option for Consumer and Producer (#149)
Add new keyword 'connection' to both
Consumer's and Producer's inits.

'connection' accepts the same values as
AMQP-library's Connection.open/2:
https://hexdocs.pm/amqp/AMQP.Connection.html#open/2

The behavior is the same as with the
'uri' keyword, but let's the user know that other
values can be used. No actual logic changes
are made. The old 'uri' keyword can still be used.